🚚 Rename to Locationcard
