+++ src/app/components/LocationCard/LocationCard.tsx
@@ -1,10 +1,6 @@
 import React, { useState } from "react";
-import styles from "./Methode.module.css";
+import styles from "./LocationCard.module.css";
 
-const locationSort = {
-  AkLocation: { name: "Aktueller Standort" },
-  value: { name: "value" },
-};
 type LabeledInputProps = {
   label: string;
   placeholder: string;
@@ -23,9 +19,6 @@ function LocationCard({
   onChange,
 }: LabeledInputProps): JSX.Element {
   const [open, setOpen] = useState(false);
-  const [locationValue, setLocationValue] = useState<"AkLocation">(
-    "AkLocation"
-  );
   if (!onChange) {
     return <div>not found.</div>;
   }
@@ -37,7 +30,7 @@ function LocationCard({
       >
         <p>Suchmethode</p>
         <label className={styles.label}>
-          {locationSort[locationValue].name}
+          {value ? value : "Aktueller Standort"}
         </label>
       </div>
       {open && (
@@ -48,7 +41,8 @@ function LocationCard({
             id="AkLocation"
             name="filter"
             value="AkLocation"
-            onChange={() => setLocationValue("AkLocation")}
+            checked={!value}
+            onChange={() => onChange("")}
           />
           <label className={styles.container__label}>{label}</label>
           <input
